requirements.txt
github3.py
launchpadlib
termcolor
